confirmation

+++ view/typeingredient/detail.php
@@ -45,7 +45,7 @@
             <div id="boutons">
                 <ul>
                     <li class="case"><?php echo '<a href="index.php?action=update&controller=typeingredient&idTypeIngredient=' . rawurlencode($type->getIdTypeIngredient()) . '"> Modifier le type</a>'; ?></li>
-                    <li class="case"><?php echo '<a href="index.php?action=delete&controller=typeingredient&idTypeIngredient=' . rawurlencode($type->getIdTypeIngredient()) . '"> Supprimer le type</a></br>';?></li>
+                    <li class="case"><?php echo '<a href="index.php?action=delete&controller=typeingredient&idTypeIngredient=' . rawurlencode($type->getIdTypeIngredient()) . '"onclick="return confirm(\'Voulez vous supprimer cette recette?\')"> Supprimer le type</a></br>';?></li>
                 </ul>
             </div>
     </div>
